Move parts of advanced settings into openvpn and wireguard settings
