+++ gui/src/renderer/lib/routes.ts
@@ -11,7 +11,9 @@ export enum RoutePath {
   accountSettings = '/settings/account',
   preferences = '/settings/preferences',
   advancedSettings = '/settings/advanced',
-  wireguardKeys = '/settings/advanced/wireguard-keys',
+  wireguardSettings = '/settings/advanced/wireguard',
+  wireguardKeys = '/settings/advanced/wireguard/keys',
+  openVpnSettings = '/settings/advanced/openvpn',
   splitTunneling = '/settings/advanced/split-tunneling',
   support = '/settings/support',
   selectLocation = '/select-location',
