+++ gui/src/shared/localization-contexts.ts
@@ -25,7 +25,10 @@ export type LocalizationContexts =
   | 'preferences-nav'
   | 'advanced-settings-view'
   | 'advanced-settings-nav'
-  | 'advanced-settings-view-wireguard'
+  | 'wireguard-settings-view'
+  | 'wireguard-settings-nav'
+  | 'openvpn-settings-view'
+  | 'openvpn-settings-nav'
   | 'wireguard-key-view'
   | 'wireguard-keys-nav'
   | 'split-tunneling-view'
